# api/openapi.yaml
openapi: 3.0.3
info:
  title: EgressDeck Controller HTTP API
  version: 0.1.0
  description: Current Go HTTP contract. Snake_case wire values are mapped by the Vue client into UI summaries.
    Collection pages use items, limit, total and optional next_cursor; the durable event feed uses its own resume
    cursor. Conditional remote adapter routes return typed HTTP 501 when unavailable. A 202 identifies an existing
    controller journal entry and is separate from applied, observed and traffic-verified state. Roadmap-only operations
    are outside paths under x-planned-operations; OIDC callback registration follows exact OIDC_REDIRECT_URL under
    x-oidc-callback.
servers:
- url: /
security:
- sessionCookie: []
...
paths:
...
  /api/v1/capabilities:
    get:
      tags:
      - gateways
      summary: Read configured controller capabilities
      operationId: getCapabilities
      responses:
        '200':
          description: Capability flags
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ControllerCapabilities'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
      x-implementation-status: implemented
      description: Controller flags describe configured service availability. events.resume=true reflects the local
        persisted polling feed. Diagnostics still read live gateway capabilities on each request. policy.apply_generation
        is an explicit qualification flag; the presence of an executor alone does not enable it. Provider/selection
        flags do not establish real-network acceptance.
      x-required-role: viewer
...
  /api/v1/providers/{id}/stage:
    parameters:
    - $ref: '#/components/parameters/ResourceID'
    post:
      tags:
      - providers
      summary: Stage supplied provider content
      operationId: stageProvider
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/CSRF'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StageProviderRequest'
      responses:
        '202':
          description: Journal operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationAccepted'
        '404':
          $ref: '#/components/responses/Error404'
        '422':
          $ref: '#/components/responses/Error422'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '400':
          description: Invalid JSON, unknown fields or request validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-implementation-status: implemented
      description: Local parse/stage only; does not publish to gateway. Records a durable operation when service
        persistence is configured.
      x-required-role: admin
...
  /api/v1/gateways/{id}/capabilities:
    parameters:
    - $ref: '#/components/parameters/ResourceID'
    get:
      tags:
      - gateways
      summary: Observe gateway health and capabilities
      operationId: getGatewayCapabilities
      responses:
        '200':
          description: Gateway view
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gateway'
        '404':
          $ref: '#/components/responses/Error404'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
      x-implementation-status: implemented
      x-required-role: viewer
...
  /api/v1/events:
    get:
      tags:
      - operations
      summary: Poll resumable management events
      operationId: getEvents
      description: Returns a bounded projection of the latest polled operation/audit state, not every transient
        transition. Contains no raw adapter payloads, actors or secrets. Retains up to 5000 projected events; durable
        indicates configured persistence. This JSON polling feed does not use SSE or Last-Event-ID. A 410 means
        reload current state and resume without the stale cursor.
      responses:
        '200':
          description: Projected event page, including a next_cursor even at an empty tail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ControllerEvents'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '400':
          description: invalid_query, invalid_limit or invalid_cursor (including a cursor ahead of the stream)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '410':
          description: cursor_expired or cursor_stream_changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: 'events_unavailable: event source or persistence unavailable'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-implementation-status: implemented
      parameters:
      - name: cursor
        in: query
        schema:
          type: string
          maxLength: 256
        description: Opaque next_cursor from a prior response; absent reads retained history.
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 500
          default: 100
      x-required-role: viewer
...
  /api/v1/gateways/{id}/connections:
    parameters:
    - $ref: '#/components/parameters/ResourceID'
    get:
      summary: getGatewayConnections
      operationId: getGatewayConnections
      tags:
      - gateways
      x-implementation-status: implemented
      x-required-capability: connections.observe
      description: Checks live gateway capabilities on every request. Query parameters and browser-selected probe
        URLs are rejected. Results are gateway-origin observations, not client traffic verification.
      responses:
        '200':
          description: Gateway diagnostic observation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayConnectionsResponse'
        '400':
          description: Query parameters or nonempty probe body rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Gateway or scoped target not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '501':
          description: Gateway diagnostic capability is unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnsupportedResponse'
        '502':
          description: invalid_gateway_response or gateway_unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
      x-required-role: viewer
  /api/v1/gateways/{id}/counters:
    parameters:
    - $ref: '#/components/parameters/ResourceID'
    get:
      summary: getGatewayCounters
      operationId: getGatewayCounters
      tags:
      - gateways
      x-implementation-status: implemented
      description: Checks live gateway capabilities on every request. Query parameters and browser-selected probe
        URLs are rejected. Results are gateway-origin observations, not client traffic verification. At least one
        of traffic.proxy_counters/traffic.direct_counters must be available. Each class has its own availability;
        provider_quota is always unavailable.
      responses:
        '200':
          description: Gateway diagnostic observation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayCountersResponse'
        '400':
          description: Query parameters or nonempty probe body rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Gateway or scoped target not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '501':
          description: Gateway diagnostic capability is unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnsupportedResponse'
        '502':
          description: invalid_gateway_response or gateway_unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
      x-required-any-capabilities:
      - traffic.proxy_counters
      - traffic.direct_counters
      x-required-role: viewer
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: egressdeck_session
      description: HTTP-only session; mutations also require X-CSRF-Token matching egressdeck_csrf.
...
  parameters:
    ResourceID:
      name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
      description: Optional controller journal key. Reuse only for the same action/payload. No HTTP-layer replay
        guarantee is provided by the gateway agent.
...
    CSRF:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
      description: Required with a session in session/header mode. Must match egressdeck_csrf cookie and session
        token.
...
  responses:
...
    Error401:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error403:
      description: Forbidden or CSRF failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error404:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Error422:
      description: Semantic validation failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      oneOf:
      - type: object
        required:
        - error
        properties:
          error:
            $ref: '#/components/schemas/ErrorDetail'
      - type: object
        required:
        - error
        - message
        properties:
          error:
            type: string
          message:
            type: string
...
    ControllerCapabilities:
      type: object
      required:
      - api_version
      - capabilities
      properties:
        api_version:
          type: string
        capabilities:
          type: object
          additionalProperties:
            type: boolean
...
    StageProviderRequest:
      type: object
      required:
      - content
      properties:
        content:
          type: string
        format:
          type: string
...
    Gateway:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        endpoint:
          type: string
        adapter:
          type: string
        observed_generation:
          type: integer
          format: int64
        revision:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        health:
          type: string
        version:
          type: string
        observation_status:
          type: string
          enum:
          - unavailable
          - failed
          - observed
        capabilities:
          type: object
          properties:
            supported:
              type: array
              items:
                type: string
              nullable: true
            restrictions:
              type: object
              additionalProperties:
                type: string
          additionalProperties: false
          required:
          - supported
      additionalProperties: false
      required:
      - id
      - name
      - endpoint
      - adapter
      - observed_generation
      - revision
      - created_at
      - updated_at
      - health
      - capabilities
      - observation_status
      x-go-type: api.GatewayView
      x-source: internal/api/gateways.go
...
    OperationAccepted:
      type: object
      properties:
        operation_id:
          type: string
        status:
          type: string
          enum:
          - draft
          - validated
          - staged
          - applying
          - verifying
          - applied
          - partially_applied
          - failed
          - outcome_unknown
        revision:
          $ref: '#/components/schemas/ProviderRevision'
        changes:
          $ref: '#/components/schemas/ChangeReport'
        items:
          type: array
          items:
            $ref: '#/components/schemas/Selection'
          nullable: true
      additionalProperties: false
      required:
      - operation_id
      - status
      description: References an existing journal entry at /api/v1/operations/{id}. Optional provider/selection
        payloads may be omitted on replay. HTTP 202 can contain a settled status; read the operation instead of
        assuming asynchronous execution or traffic success.
...
    ControllerEvents:
      type: object
      additionalProperties: false
      required:
      - items
      - next_cursor
      - has_more
      - resumable
      - durable
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ManagementEvent'
        next_cursor:
          type: string
        has_more:
          type: boolean
        resumable:
          type: boolean
          enum:
          - true
        durable:
          type: boolean
      x-go-type: api.eventPage
      x-source: internal/api/events.go
...